Add write method to logger
Added a write function to the logger, to write unformatted strings
directly to the output.

Fixed the assertion macro to still evaluate the expression,  without
validating it in release mode.

Disabled the set sampling rate function in the i2s module, isnce it
wasn't working as expected (it just hangs)